Pull out middleware code into its own package
